requirements.txt
discord.py>=2.3.0
python-dotenv
markovify